refactor(configaudit): extract shared file-metadata helpers
Move ownerInfo, defaultInGitRepo, sha256Hex, and the git-tracked exec
glue into a single meta.go shared by both npmrc and pip detectors.
Drops the pipOwnerInfo / pipStatOwner / pipDefaultInGitRepo duplicates
and the two pip-only stat shims. No behavior change.
