+++ internal/detector/configaudit/npmrc.go
@@ -68,20 +68,12 @@ type NPMRCDetector struct {
 	inGitRepo func(path string) bool
 }
 
-type ownerInfo struct {
-	UID       int
-	GID       int
-	OwnerName string
-	GroupName string
-	OK        bool
-}
-
 // NewNPMRCDetector returns a detector with default platform-specific
 // metadata helpers wired in.
 func NewNPMRCDetector(exec executor.Executor) *NPMRCDetector {
 	d := &NPMRCDetector{exec: exec}
-	d.ownerLookup = func(p string) ownerInfo { return statOwner(p) }
-	d.gitTracked = d.defaultGitTracked
+	d.ownerLookup = statOwner
+	d.gitTracked = func(ctx context.Context, p string) bool { return gitTrackedViaExec(ctx, exec, p) }
 	d.inGitRepo = defaultInGitRepo
 	return d
 }
@@ -426,39 +418,3 @@ func (d *NPMRCDetector) collectEnv() []model.NPMRCEnvVar {
 	return out
 }
 
-// defaultGitTracked shells out to git to check if a file is tracked.
-// Returns false on any error (git not installed, not in a repo, untracked).
-func (d *NPMRCDetector) defaultGitTracked(ctx context.Context, path string) bool {
-	dir := filepath.Dir(path)
-	base := filepath.Base(path)
-	_, _, exit, err := d.exec.RunWithTimeout(ctx, 5*time.Second, "git", "-C", dir, "ls-files", "--error-unmatch", base)
-	return err == nil && exit == 0
-}
-
-// defaultInGitRepo walks parent directories looking for a .git entry.
-// Stops at the filesystem root.
-func defaultInGitRepo(path string) bool {
-	dir := filepath.Dir(path)
-	for {
-		gitPath := filepath.Join(dir, ".git")
-		if info, err := os.Stat(gitPath); err == nil {
-			// .git can be a directory (regular repo) or a file (worktree).
-			_ = info
-			return true
-		}
-		parent := filepath.Dir(dir)
-		if parent == dir {
-			return false
-		}
-		dir = parent
-	}
-}
-
-// sha256Hex returns the hex SHA-256 of a string.
-func sha256Hex(s string) string {
-	if s == "" {
-		return ""
-	}
-	sum := sha256.Sum256([]byte(s))
-	return hex.EncodeToString(sum[:])
-}
